Add Formatter class
Formatter class has the sole reponsibility of processing a List of persons
and into the text UI.

This is done by taking each element (String) in the list and adding a
number prefix and a newline character postfix.

The returning object is a nicely formatted String that is returned back to
TextUi to be printed.
